// frontend/src/client/types.gen.ts
export type CollectionCreate = {
    name: string;
    description?: (string | null);
    is_public?: boolean;
};
// ...
export type UserPublic = {
    email: string;
    is_active?: boolean;
    is_superuser?: boolean;
    full_name?: (string | null);
    id: number;
};